+++ hug/interface.py
@@ -48,6 +48,12 @@ from hug.types import (
     text,
 )
 
+DOC_TYPE_MAP = {str: "String", bool: "Boolean", list: "Multiple", int: "Integer", float: "Float"}
+
+
+def _doc(kind):
+    return DOC_TYPE_MAP.get(kind, kind.__doc__)
+
 
 def asyncio_call(function, *args, **kwargs):
     loop = asyncio.get_event_loop()
@@ -227,7 +233,7 @@ class Interface(object):
             self.output_doc = self.transform.__doc__
         elif self.transform or self.interface.transform:
             output_doc = self.transform or self.interface.transform
-            self.output_doc = output_doc if type(output_doc) is str else output_doc.__doc__
+            self.output_doc = output_doc if type(output_doc) is str else _doc(output_doc)
 
         self.raise_on_invalid = route.get("raise_on_invalid", False)
         if "on_invalid" in route:
@@ -310,7 +316,7 @@ class Interface(object):
                 for requirement in self.requires
             ]
         doc["outputs"] = OrderedDict()
-        doc["outputs"]["format"] = self.outputs.__doc__
+        doc["outputs"]["format"] = _doc(self.outputs)
         doc["outputs"]["content_type"] = self.outputs.content_type
         parameters = [
             param
@@ -329,7 +335,7 @@ class Interface(object):
                     continue
 
                 input_definition = inputs.setdefault(argument, OrderedDict())
-                input_definition["type"] = kind if isinstance(kind, str) else kind.__doc__
+                input_definition["type"] = kind if isinstance(kind, str) else _doc(kind)
                 default = self.defaults.get(argument, None)
                 if default is not None:
                     input_definition["default"] = default
@@ -505,7 +511,7 @@ class CLI(Interface):
             if option in self.interface.input_transformations:
                 transform = self.interface.input_transformations[option]
                 kwargs["type"] = transform
-                kwargs["help"] = transform.__doc__
+                kwargs["help"] = _doc(transform)
                 if transform in (list, tuple) or isinstance(transform, types.Multiple):
                     kwargs["action"] = "append"
                     kwargs["type"] = Text()
@@ -588,9 +594,12 @@ class CLI(Interface):
 
         for field, type_handler in self.reaffirm_types.items():
             if field in pass_to_function:
-                pass_to_function[field] = self.initialize_handler(
-                    type_handler, pass_to_function[field], context=context
-                )
+                if not pass_to_function[field] and type_handler in (list, tuple, hug.types.Multiple):
+                    pass_to_function[field] = type_handler(())
+                else:
+                    pass_to_function[field] = self.initialize_handler(
+                        type_handler, pass_to_function[field], context=context
+                    )
 
         if getattr(self, "validate_function", False):
             errors = self.validate_function(pass_to_function)