+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Ideally, within a virtual environment.
 
 Changelog
 =========
+### 2.6.0 - August 28, 2019
+- Improved CLI multiple behaviour with empty defaults
+- Improved CLI type output for built-in types
+
 ### 2.5.6 - June 20, 2019
 - Fixed issue #815: map_params() causes api documentation to lose param type information
 - Improved project testing: restoring 100% coverage